tests/requirements.txt
pytest
pytest-asyncio
pytest-html
PyHamcrest
starlette
httpx
websockets
pydantic